refactor: use middleware method to validate date, extract data prensentation in StatAction
